update package names for consistency
