+++ best-practice/src/test/java/com/saucedemo/pages/LoginPage.java
@@ -1,4 +1,4 @@
-package com.pages;
+package com.saucedemo.pages;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
