+++ best-practice/src/test/java/com/saucedemo/pages/ShoppingCartPage.java
@@ -1,4 +1,4 @@
-package com.pages;
+package com.saucedemo.pages;
 
 import org.openqa.selenium.WebDriver;
