+++ best-practice/src/test/java/com/saucedemo/tests/DesktopTests.java
@@ -1,7 +1,7 @@
 package com.saucedemo.tests;
 
-import com.pages.LoginPage;
-import com.pages.ProductsPage;
+import com.saucedemo.pages.LoginPage;
+import com.saucedemo.pages.ProductsPage;
 import com.saucedemo.WebTestsBase;
 import com.saucelabs.saucebindings.SauceOptions;
 import com.saucelabs.saucebindings.SauceSession;
