+++ best-practice/src/test/java/com/saucedemo/tests/EmusimWebAppTests.java
@@ -1,7 +1,7 @@
 package com.saucedemo.tests;
 
-import com.pages.LoginPage;
-import com.pages.ProductsPage;
+import com.saucedemo.pages.LoginPage;
+import com.saucedemo.pages.ProductsPage;
 import com.saucedemo.Endpoints;
 import com.saucedemo.WebTestsBase;
 import org.junit.Before;
