+++ selenium-junit4-examples/src/test/java/com/saucedemo/selenium/junit4/SauceConnectTest.java
@@ -1,4 +1,4 @@
-package com.saucedemo;
+package com.saucedemo.selenium.junit4;
 
 import com.saucelabs.saucebindings.SauceOptions;
 import com.saucelabs.saucebindings.SauceSession;
