+++ selenium-junit4-examples/src/test/java/com/saucedemo/selenium/junit4/WindowsAuthentication.java
@@ -1,4 +1,4 @@
-package com.saucedemo;
+package com.saucedemo.selenium.junit4;
 
 import com.saucelabs.saucebindings.Prerun;
 import com.saucelabs.saucebindings.SauceOptions;
