+++ selenium-testng-examples/src/test/java/com/saucedemo/selenium/testng/ParallelSingleBrowserTest.java
@@ -1,4 +1,4 @@
-package com.saucedemo;
+package com.saucedemo.selenium.testng;
 
 import com.saucelabs.saucebindings.SauceOptions;
 import com.saucelabs.saucebindings.SauceSession;
